wip: disable array and object
