+++ packages/sanity/src/core/field/diff/resolve/defaultComponents.ts
@@ -7,17 +7,17 @@ import {NumberFieldDiff} from '../../types/number/diff'
 import {PTDiff} from '../../types/portableText/diff'
 import {ReferenceFieldDiff} from '../../types/reference/diff'
 import {StringFieldDiff} from '../../types/string/diff'
-import {ObjectFieldDiff} from '../../types/object/diff'
-import {ArrayFieldDiff} from '../../types/array/diff/ArrayFieldDiff'
+// import {ObjectFieldDiff} from '../../types/object/diff'
+// import {ArrayFieldDiff} from '../../types/array/diff/ArrayFieldDiff'
 
 export const defaultComponents: Record<string, ComponentType<any> | undefined> = {
-  array: ArrayFieldDiff,
+  // array: ArrayFieldDiff, // TODO
   block: PTDiff,
   boolean: BooleanFieldDiff,
   file: FileFieldDiff,
   image: ImageFieldDiff,
   number: NumberFieldDiff,
-  object: ObjectFieldDiff,
+  // object: ObjectFieldDiff, // TODO
   reference: ReferenceFieldDiff,
   string: StringFieldDiff,
 }
